requirements.txt
python-dotenv
langchain
openai
streamlit
langchain-openai  